Add files via upload
Animation gif of cuttlefish landings in the English Channel by month from 2000 to 2022. The size of the piecharts is proportional to the total landings for each ICES rectangle. The colors of the piecharts represent the share of landings captured by the French fleet (in blue) and by the UK fleet (in red). This is a series of 276 maps that appear in sequence.